chore(product tours): move URL targeting to auto-show conditions
